Refactor: split out tests for `complexities` module
